Rename LobbyScreen to WelcomeScreen and finalize navigation flow
