+++ app/src/main/java/com/doomhamsters/MainActivity.kt
@@ -1,40 +1,16 @@
 package com.doomhamsters
-
-import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.doomhamsters.ui.theme.DoomHamstersTheme
 import androidx.lifecycle.viewmodel.compose.viewModel
 
-import com.doomhamsters.ui.theme.Orange
-import com.doomhamsters.ui.theme.SoftWhite
-import com.doomhamsters.ui.theme.WarmAlmond
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -46,74 +22,17 @@ class MainActivity : ComponentActivity() {
                 val lobbyViewModel: LobbyViewModel = viewModel()
                 Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
                     // 3. Lobby Steuerung
-                    Box(modifier = Modifier.padding(innerPadding)) {
+                    Box(modifier = Modifier
+                        .padding(innerPadding)) {
 
                         // Navigation
                         MainLobbyNavigation(lobbyViewModel)
 
-                        Greeting(
-                            name = "Android",
-                            modifier = Modifier.padding(innerPadding)
-                        )
+
 
                     }
                 }
             }
         }
-    }
-
-@Composable
-fun LobbyScreen(modifier: Modifier = Modifier) {
-    val context = LocalContext.current
-
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .background(WarmAlmond)
-            .padding(24.dp),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(
-            text = stringResource(R.string.welcome_message),
-            textAlign = TextAlign.Center,
-            lineHeight = 20.sp,
-            fontSize = 32.sp,
-            fontWeight = FontWeight.Bold,
-            color = Orange
-        )
-
-        Spacer(modifier = Modifier.height(40.dp))
-
-        Button(
-            onClick = {
-                val intent = Intent(context, RulesActivity::class.java)
-                context.startActivity(intent)
-            },
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(56.dp),
-            shape = RoundedCornerShape(50),
-            colors = ButtonDefaults.buttonColors(
-                containerColor = Orange,
-                contentColor = SoftWhite
-            ),
-            elevation = ButtonDefaults.buttonElevation(defaultElevation = 3.dp)
-        ) {
-            Text(
-                text = stringResource(R.string.rules_button_text),
-                fontSize = 18.sp,
-                fontWeight = FontWeight.Bold,
-                color = SoftWhite
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun LobbyScreenPreview() {
-    DoomHamstersTheme {
-        LobbyScreen()
     }
 }
